# 24001200-km7-afi-testingAuth-ch5/docs/swagger.yml
openapi: 3.0.0
info:
  title: Banking System API
  version: 1.0.0
  description: API documentation for Banking System

servers:
  - url: http://localhost:3000

paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        201:
          description: User registered successfully
        400:
          description: Email already exists
        500:
          description: Registration failed

  /api/v1/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
        400:
          description: Invalid credentials
        404:
          description: User not found
        500:
          description: Login failed

  /api/v1/auth/authenticate:
    get:
      summary: Authenticate token
      security:
        - bearerAuth: []
      responses:
        200:
          description: Authentication successful
        401:
          description: Invalid token

  /api/v1/users:
    post:
      summary: Create a new user with a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                profile:
                  type: object
                  properties:
                    identity_type:
                      type: string
                    identity_number:
                      type: string
                    address:
                      type: string
      responses:
        '201':
          description: User created successfully
        '500':
          description: Error creating user

    get:
      summary: Get a list of all users
      responses:
        '200':
          description: Successfully fetched users
        '500':
          description: Error fetching users

  /api/v1/users/{userId}:
    get:
      summary: Get user details by ID
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: ID of the user
      responses:
        '200':
          description: Successfully fetched user details
        '404':
          description: User not found
        '500':
          description: Error fetching user

    put:
      summary: Update user details by ID
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully updated user
        '404':
          description: User not found
        '500':
          description: Error updating user

  /api/v1/accounts:
    post:
      summary: Create a new bank account for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                bank_name:
                  type: string
                bank_account_number:
                  type: string
                balance:
                  type: number
      responses:
        '201':
          description: Bank account created successfully
        '500':
          description: Error creating account

    get:
      summary: Get a list of all bank accounts
      responses:
        '200':
          description: Successfully fetched accounts
        '500':
          description: Error fetching accounts

  /api/v1/accounts/user/{userId}:
    get:
      summary: Get all bank accounts by user ID
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: ID of the user
      responses:
        '200':
          description: Successfully fetched accounts
        '404':
          description: No accounts found for this user
        '500':
          description: Error fetching accounts

  /api/v1/transactions:
    post:
      summary: Transfer money between two accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sourceAccountId:
                  type: integer
                destinationAccountId:
                  type: integer
                amount:
                  type: number
      responses:
        '200':
          description: Transaction created successfully
        '404':
          description: Source or destination account not found
        '500':
          description: Error creating transaction

    get:
      summary: Get a list of all transactions
      responses:
        '200':
          description: Successfully fetched transactions
        '500':
          description: Error fetching transactions

  /api/v1/transactions/{transactionId}:
    get:
      summary: Get transaction details by ID
      parameters:
        - in: path
          name: transactionId
          schema:
            type: integer
          required: true
          description: ID of the transaction
      responses:
        '200':
          description: Successfully fetched transaction details
        '404':
          description: Transaction not found
        '500':
          description: Error fetching transaction

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT